Adds controllersContacts
